+++ src/combinator/ops/mul/fold.rs
@@ -72,9 +72,9 @@ mod tests {
 
   #[test]
   fn combinator_mul_usize() {
-    let rejecter = || wrap(|_, _| Option::<Output<()>>::None);
+    let rejecter = || wrap(|_| Option::<Output<()>>::None);
     let accepter = || {
-      wrap(|instant, _| {
+      wrap(|instant| {
         instant
           .accept(1)
           .map(|output| output.map(|_| instant.digested()))
@@ -122,9 +122,9 @@ mod tests {
 
   #[test]
   fn combinator_mul_range() {
-    let rejecter = || wrap(|_, _| Option::<Output<()>>::None);
+    let rejecter = || wrap(|_| Option::<Output<()>>::None);
     let accepter = || {
-      wrap(|instant, _| {
+      wrap(|instant| {
         instant
           .accept(1)
           .map(|output| output.map(|_| instant.digested()))
@@ -174,9 +174,9 @@ mod tests {
 
   #[test]
   fn combinator_mul_range_from() {
-    let rejecter = || wrap(|_, _| Option::<Output<()>>::None);
+    let rejecter = || wrap(|_| Option::<Output<()>>::None);
     let accepter = || {
-      wrap(|instant, _| {
+      wrap(|instant| {
         instant
           .accept(1)
           .map(|output| output.map(|_| instant.digested()))
@@ -216,9 +216,9 @@ mod tests {
 
   #[test]
   fn combinator_mul_range_full() {
-    let rejecter = || wrap(|_, _| Option::<Output<()>>::None);
+    let rejecter = || wrap(|_| Option::<Output<()>>::None);
     let accepter = || {
-      wrap(|instant, _| {
+      wrap(|instant| {
         instant
           .accept(1)
           .map(|output| output.map(|_| instant.digested()))
@@ -248,9 +248,9 @@ mod tests {
 
   #[test]
   fn combinator_mul_range_inclusive() {
-    let rejecter = || wrap(|_, _| Option::<Output<()>>::None);
+    let rejecter = || wrap(|_| Option::<Output<()>>::None);
     let accepter = || {
-      wrap(|instant, _| {
+      wrap(|instant| {
         instant
           .accept(1)
           .map(|output| output.map(|_| instant.digested()))
@@ -300,9 +300,9 @@ mod tests {
 
   #[test]
   fn combinator_mul_range_to() {
-    let rejecter = || wrap(|_, _| Option::<Output<()>>::None);
+    let rejecter = || wrap(|_| Option::<Output<()>>::None);
     let accepter = || {
-      wrap(|instant, _| {
+      wrap(|instant| {
         instant
           .accept(1)
           .map(|output| output.map(|_| instant.digested()))
@@ -342,9 +342,9 @@ mod tests {
 
   #[test]
   fn combinator_mul_range_to_inclusive() {
-    let rejecter = || wrap(|_, _| Option::<Output<()>>::None);
+    let rejecter = || wrap(|_| Option::<Output<()>>::None);
     let accepter = || {
-      wrap(|instant, _| {
+      wrap(|instant| {
         instant
           .accept(1)
           .map(|output| output.map(|_| instant.digested()))